Codename "Cerberus": initial commit
This is just a very bare copy of the ARC++ source tree for now. It's
going to be quite a wondrous recode. We're basically doing onlink from
scratch, without any of the Introversion code at all, which allows us to
do lots of things we never could do before. For instance, make the
project open source. But that'll be later. For now, let's just get some
prototypes!

Signed-off-by: Steven Noonan <steven@uplinklabs.net>